apps/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "infer type of csv and excel files"
authors = ["ducduyn31 <ducduy.ababat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.3"
pandas = "^2.2.1"
psycopg = "^3.1.18"
gunicorn = "^21.2.0"
djangorestframework = "^3.15.1"
minio = "^7.2.5"
transitions = "^0.9.0"
django-ulid = "^0.0.4"
django-auditlog = "^2.3.0"
celery = {extras = ["librabbitmq", "msgpack"], version = "^5.3.6"}
uvicorn = "^0.29.0"
supervisor = "^4.2.5"
python-magic = "^0.4.27"
boto3 = "^1.34.76"
django-eventstream = "^5.1.0"


[tool.poetry.group.dev.dependencies]
celery = {extras = ["pytest"], version = "^5.3.6"}
pytest = "^8.1.1"
requests = "^2.31.0"
faker = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
